refactor: split semantic-svc app.py into focused modules (ADR-0037)
Splits the 1248-line semantic-svc/app.py monolith into six focused
modules per ADR-0037:

- models.py (97 lines): All 16 Pydantic request/response schemas
- retention.py (228 lines): Domain category classification,
  retention scoring, and capacity eviction
- router_index.py (349 lines): /index endpoints + payload builder
- router_search.py (69 lines): /search/vector endpoint
- router_migration.py (244 lines): /index/migrate/* endpoints
- app.py (410 lines, down from 1248): App factory, lifespan,
  model helpers, middleware, /health, /embed, /rerank, /metrics

Uses FastAPI APIRouter pattern with prefix mounting. All URL paths,
response schemas, and HTTP status codes remain identical. Dockerfile
updated to include new module files.

See ADR-0037 for the full design rationale and module boundaries.
Follows the same architectural pattern as ADR-0036 (fetch.py split).

Closes #244
