+++ semantic-svc/Dockerfile
@@ -7,7 +7,7 @@ RUN pip install --no-cache-dir --extra-index-url https://download.pytorch.org/wh
 # Pre-download embedding model at build time so first request is fast
 RUN python -c "from sentence_transformers import SentenceTransformer; SentenceTransformer('BAAI/bge-m3')"
 
-COPY app.py metrics.py ./
+COPY app.py models.py retention.py router_index.py router_search.py router_migration.py metrics.py ./
 
 EXPOSE 8003
 CMD ["uvicorn", "app:app", "--host", "0.0.0.0", "--port", "8003"]
