+++ semantic-svc/app.py
@@ -27,20 +27,24 @@ Phase 4 endpoints (this file):
 import asyncio
 import datetime
 import hashlib
-import json
 import logging
-import math
 import os
 import time
-import urllib.parse
 from contextlib import asynccontextmanager
 
 import numpy as np
 from fastapi import FastAPI, HTTPException, Request, Response
-from metrics import METRICS
-from pydantic import BaseModel
 from qdrant_client import QdrantClient, models
 from sentence_transformers import CrossEncoder, SentenceTransformer
+
+from metrics import METRICS
+from models import (
+    EmbedRequest,
+    EmbedResponse,
+    RerankRequest,
+    RerankResult,
+    RerankResponse,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -107,7 +111,7 @@ async def lifespan(app: FastAPI):
         )
         _models_ready = True
         logger.info("Models loaded — semantic-svc ready")
-    except (OSError, RuntimeError, ValueError):
+    except Exception:
         logger.exception(
             "Failed to load semantic models — /health will report 'starting'"
         )
@@ -171,146 +175,16 @@ def _get_active_model() -> str:
     return _active_override if _active_override is not None else ACTIVE_EMBED_MODEL
 
 
-# ── Domain classification ─────────────────────────────────────────
-
-# Domain patterns for retention categories (unchanged from Phase 3)
-_DOCS_DOMAINS = {
-    "readthedocs.io",
-    "readthedocs.org",
-}
-_REFERENCE_DOMAINS = {
-    "wikipedia.org",
-    "stackoverflow.com",
-    "stackexchange.com",
-    "github.com",
-    "gitlab.com",
-}
-_NEWS_DOMAINS = {
-    "reuters.com",
-    "nytimes.com",
-    "cnn.com",
-    "bbc.com",
-    "bbc.co.uk",
-    "bloomberg.com",
-    "apnews.com",
-    "npr.org",
-    "theguardian.com",
-    "wsj.com",
-    "washingtonpost.com",
-    "economist.com",
-    "cnbc.com",
-    "abcnews.go.com",
-    "cbsnews.com",
-    "nbcnews.com",
-    "usatoday.com",
-    "politico.com",
-    "axios.com",
-    "thehill.com",
-}
-_SOCIAL_DOMAINS = {
-    "reddit.com",
-    "twitter.com",
-    "x.com",
-    "youtube.com",
-    "instagram.com",
-    "tiktok.com",
-    "threads.net",
-    "bluesky",
-    "bsky.app",
-}
-_BLOG_DOMAINS = {
-    "medium.com",
-    "substack.com",
-    "ghost.io",
-    "wordpress.com",
-    "blogger.com",
-    "blogspot.com",
-}
+def _set_active_override(val: str | None) -> None:
+    """Set the in-memory active model override (used by migration cutover)."""
+    global _active_override
+    _active_override = val
 
 
-def _compute_domain_category(url: str) -> str:
-    """Classify a URL's domain into a retention category.
-
-    Categories (in eviction-priority order):
-        news (0.3), social (0.4), blog (0.6), api (0.7),
-        unknown (0.8), reference (1.0), docs (1.2)
-    """
-    netloc = urllib.parse.urlparse(url).netloc.lower()
-
-    if not netloc:
-        return "unknown"
-
-    # Strip leading www. for consistent matching
-    netloc = netloc.removeprefix("www.")
-
-    # Prefix-based: docs./learn./help./api./developer.
-    if netloc.startswith(("docs.", "learn.", "help.")):
-        return "docs"
-    if netloc.startswith(("api.", "developer.")):
-        return "api"
-    if netloc.startswith("blog."):
-        return "blog"
-
-    # Substring-based matches
-    for d in _DOCS_DOMAINS:
-        if d in netloc:
-            return "docs"
-    for d in _REFERENCE_DOMAINS:
-        if d in netloc:
-            return "reference"
-    for d in _NEWS_DOMAINS:
-        if d in netloc:
-            return "news"
-    for d in _SOCIAL_DOMAINS:
-        if d in netloc:
-            return "social"
-    for d in _BLOG_DOMAINS:
-        if d in netloc:
-            return "blog"
-
-    return "unknown"
-
-
-# ── Retention scoring ─────────────────────────────────────────────
-
-_DOMAIN_MULTIPLIERS = {
-    "news": 0.3,
-    "social": 0.4,
-    "blog": 0.6,
-    "api": 0.7,
-    "unknown": 0.8,
-    "reference": 1.0,
-    "docs": 1.2,
-}
-
-_RECENCY_HALF_LIFE_DAYS = 90.0
-
-
-def _compute_retention_score(payload: dict) -> float:
-    """Compute retention score from a Qdrant point's payload."""
-    category = payload.get("domain_category", "unknown")
-    domain_mult = _DOMAIN_MULTIPLIERS.get(category, 0.8)
-
-    raw_date = payload.get("last_indexed_at", "")
-    if raw_date:
-        try:
-            last_indexed = datetime.datetime.fromisoformat(raw_date)
-            days_since = (datetime.datetime.now(datetime.UTC) - last_indexed).days
-            days_since = max(0, days_since)
-            recency_factor = math.exp(-days_since / _RECENCY_HALF_LIFE_DAYS)
-            recency_factor = max(0.1, min(1.0, recency_factor))
-        except (ValueError, TypeError):
-            recency_factor = 0.5
-    else:
-        recency_factor = 0.5
-
-    access_count = payload.get("access_count", 0)
-    access_boost = min(int(access_count), 100) * 0.01
-
-    crawl_count = payload.get("crawl_count", 0)
-    crawl_boost = min(int(crawl_count), 20) * 0.05
-
-    return round(domain_mult * recency_factor + access_boost + crawl_boost, 4)
+def _set_migration_task(task: asyncio.Task | None) -> None:
+    """Set the migration background task reference."""
+    global _migration_task
+    _migration_task = task
 
 
 # ── Model helpers ─────────────────────────────────────────────────
@@ -429,339 +303,6 @@ async def _ensure_qdrant() -> QdrantClient:
     return _qdrant
 
 
-# ── Scoring-based eviction (Phase 3) ──────────────────────────────
-
-
-async def _evict_if_needed(qdrant: QdrantClient):
-    """Scoring-based eviction if the index exceeds MAX_DOCS."""
-    count = qdrant.count(COLLECTION_NAME).count
-    if count <= MAX_DOCS:
-        return
-
-    excess = count - MAX_DOCS
-    target_delete = excess + max(100, int(MAX_DOCS * 0.02))
-
-    logger.info(
-        "Index over capacity (%d / %d), scoring eviction candidates...",
-        count,
-        MAX_DOCS,
-    )
-
-    scored_points: list[tuple[float, int]] = []
-    next_offset: int | None = None
-    page_size = 2000
-
-    while len(scored_points) < target_delete or next_offset is not None:
-        page, next_offset = qdrant.scroll(
-            COLLECTION_NAME,
-            limit=page_size,
-            offset=next_offset,
-            with_payload=True,
-            with_vectors=False,
-        )
-        if not page:
-            break
-        for point in page:
-            if point.payload is None:
-                continue
-            score = _compute_retention_score(point.payload)
-            scored_points.append((score, point.id))
-
-        if next_offset is None:
-            break
-
-    scored_points.sort(key=lambda x: x[0])
-
-    to_delete = [pid for _, pid in scored_points[:target_delete]]
-
-    if to_delete:
-        qdrant.delete(
-            COLLECTION_NAME,
-            points_selector=models.PointIdsList(points=to_delete),
-        )
-        new_count = qdrant.count(COLLECTION_NAME).count
-        METRICS.counter(
-            "groktocrawl_index_evictions_total",
-            "Cumulative evictions from the vector index",
-        ).inc(value=len(to_delete))
-        logger.info(
-            "Scored eviction: removed %d documents (score range: %.4f – %.4f). "
-            "Index at %d / %d",
-            len(to_delete),
-            scored_points[0][0] if scored_points else 0,
-            scored_points[min(len(scored_points), target_delete) - 1][0]
-            if scored_points
-            else 0,
-            new_count,
-            MAX_DOCS,
-        )
-    else:
-        logger.info("No eviction candidates found")
-
-
-# ── Migration logic ────────────────────────────────────────────────
-
-
-async def _run_backfill(qdrant: QdrantClient, target_name: str, target_dim: int):
-    """Background task: scroll all points, re-embed with new model, add named vector.
-
-    This runs as an asyncio task and updates _migration state as it progresses.
-    """
-    global _migration
-    logger.info(
-        "Migration backfill started: %s (%d) -> %s (%d)",
-        EMBED_MODEL_NAME,
-        EMBED_DIM,
-        target_name,
-        target_dim,
-    )
-
-    # Load the target embedding model
-    target_model = SentenceTransformer(target_name)
-
-    # Count total docs and scroll
-    total = qdrant.count(COLLECTION_NAME).count
-    _migration["docs_total"] = total
-    _migration["status"] = "backfilling"
-
-    processed = 0
-    next_offset: int | None = None
-    page_size = 100  # Smaller batch to avoid OOM on large embeddings
-
-    try:
-        while True:
-            page, next_offset = qdrant.scroll(
-                COLLECTION_NAME,
-                limit=page_size,
-                offset=next_offset,
-                with_payload=True,
-                with_vectors=False,
-            )
-            if not page:
-                break
-
-            nv_name = _named_vector_name(target_name)
-            now = _now_iso()
-
-            for point in page:
-                if point.payload is None:
-                    continue
-                content = (
-                    point.payload.get("title", "") + " " + point.payload.get("url", "")
-                )
-                # If we had original content stored, we'd use that. For backfill,
-                # we embed from URL+title as a stopgap. Full re-indexing would
-                # re-scrape the source.
-                # Embed with the new model
-                try:
-                    # Use a truncated content signal for backfill embedding
-                    signal = f"{point.payload.get('title', '')} {point.payload.get('url', '')}"
-                    if not signal.strip():
-                        signal = point.payload.get("url", "")
-                    embedding = target_model.encode(
-                        signal[:2000], normalize_embeddings=True
-                    ).tolist()
-                except Exception as e:
-                    logger.warning(
-                        "Backfill embed failed for point %s: %s", point.id, e
-                    )
-                    continue
-
-                # Update the point: add named vector + update embedding_models list
-                existing_models = json.loads(
-                    point.payload.get("embedding_models", "[]")
-                )
-                model_short = _named_vector_name(EMBED_MODEL_NAME)
-                target_short = nv_name
-                if model_short not in existing_models:
-                    existing_models.append(model_short)
-                if target_short not in existing_models:
-                    existing_models.append(target_short)
-
-                qdrant.upsert(
-                    COLLECTION_NAME,
-                    points=[
-                        models.PointStruct(
-                            id=point.id,
-                            vector={
-                                nv_name: embedding,
-                            },
-                            payload={
-                                "embedding_model": EMBED_MODEL_NAME,
-                                "embedding_dim": EMBED_DIM,
-                                "embedding_models": existing_models,
-                            },
-                        )
-                    ],
-                )
-                processed += 1
-                if processed % 100 == 0:
-                    _migration["docs_processed"] = processed
-                    logger.info("Migration backfill: %d / %d", processed, total)
-
-            if next_offset is None:
-                break
-
-        _migration["docs_processed"] = processed
-        _migration["status"] = "dual_write"
-        logger.info(
-            "Migration backfill complete: %d / %d documents. Entering dual-write phase.",
-            processed,
-            total,
-        )
-
-    except Exception as e:
-        _migration["status"] = "idle"
-        logger.error("Migration backfill failed: %s", e)
-        raise
-
-
-# ── Request/Response models ──────────────────────────────────────
-
-
-class EmbedRequest(BaseModel):
-    model: str = "BGE-M3"
-    input: list[str]
-
-
-class EmbedResponse(BaseModel):
-    embeddings: list[list[float]]
-
-
-class RerankRequest(BaseModel):
-    query: str
-    documents: list[str]
-    top_k: int = 5
-
-
-class RerankResult(BaseModel):
-    index: int
-    score: float
-
-
-class RerankResponse(BaseModel):
-    results: list[RerankResult]
-
-
-class IndexRequest(BaseModel):
-    url: str
-    title: str = ""
-    content: str
-
-
-class IndexResponse(BaseModel):
-    status: str
-    url_hash: int
-
-
-class IndexBatchRequest(BaseModel):
-    pages: list[IndexRequest]
-
-
-class IndexBatchResponse(BaseModel):
-    status: str
-    count: int
-
-
-class VectorSearchRequest(BaseModel):
-    query: str
-    limit: int = 5
-
-
-class VectorSearchResult(BaseModel):
-    url: str
-    title: str
-    score: float
-
-
-class VectorSearchResponse(BaseModel):
-    results: list[VectorSearchResult]
-
-
-class IndexStatsResponse(BaseModel):
-    total_docs: int
-    max_docs: int
-
-
-class ModelInfoResponse(BaseModel):
-    current_model: str
-    current_dim: int
-    active_named_vector: str
-    collection: str
-    total_docs: int
-    max_docs: int
-    migration: dict
-
-
-class MigrationStartRequest(BaseModel):
-    target_model: str
-    target_dim: int
-
-
-class MigrationStatusResponse(BaseModel):
-    status: str
-    source_model: str
-    source_dim: int
-    target_model: str
-    target_dim: int
-    docs_processed: int
-    docs_total: int
-    started_at: str
-    completed_at: str
-
-
-# ── Payload building ──────────────────────────────────────────────
-
-
-def _build_index_payload(url: str, title: str, existing_payload: dict | None) -> dict:
-    """Build enriched payload for a new or updated index entry."""
-    now = _now_iso()
-    domain_category = _compute_domain_category(url)
-
-    if existing_payload:
-        first_indexed = existing_payload.get("first_indexed_at", now)
-        access_count = int(existing_payload.get("access_count", 0))
-        last_accessed = existing_payload.get("last_accessed_at", "")
-        crawl_count = int(existing_payload.get("crawl_count", 0)) + 1
-        # Preserve model metadata if re-indexing
-        embedding_model = existing_payload.get("embedding_model", EMBED_MODEL_NAME)
-        embedding_dim = int(existing_payload.get("embedding_dim", EMBED_DIM))
-        embedding_models_raw = existing_payload.get("embedding_models", "[]")
-        try:
-            embedding_models = (
-                json.loads(embedding_models_raw)
-                if isinstance(embedding_models_raw, str)
-                else list(embedding_models_raw)
-            )
-        except (json.JSONDecodeError, TypeError):
-            embedding_models = [_named_vector_name(embedding_model)]
-    else:
-        first_indexed = now
-        access_count = 0
-        last_accessed = ""
-        crawl_count = 1
-        embedding_model = EMBED_MODEL_NAME
-        embedding_dim = EMBED_DIM
-        embedding_models = [_named_vector_name(embedding_model)]
-
-    payload = {
-        "url": url,
-        "title": title,
-        "domain_category": domain_category,
-        "first_indexed_at": first_indexed,
-        "last_indexed_at": now,
-        "crawl_count": crawl_count,
-        "access_count": access_count,
-        "last_accessed_at": last_accessed,
-        "embedding_model": embedding_model,
-        "embedding_dim": embedding_dim,
-        "embedding_models": json.dumps(embedding_models),
-    }
-
-    payload["retention_score"] = _compute_retention_score(payload)
-    return payload
-
-
 # ── Metrics middleware ──────────────────────────────────────────────
 
 
@@ -790,7 +331,7 @@ async def metrics_middleware(request: Request, call_next):
         ).observe({"endpoint": path}, duration)
 
 
-# ── Endpoints ────────────────────────────────────────────────────
+# ── Core endpoints ──────────────────────────────────────────────────
 
 
 @app.get("/health")
@@ -839,402 +380,6 @@ async def rerank(body: RerankRequest):
     return RerankResponse(results=results)
 
 
-# ── Phase 2/3/4: Vector Index ────────────────────────────────────
-
-
-@app.post("/index", response_model=IndexResponse, status_code=201)
-async def index_page(body: IndexRequest):
-    """Embed and store a page in the persistent vector index.
-
-    Phase 4: Uses named vectors. During dual-write migration,
-    indexes with both the active model and the target model.
-    """
-    if not _models_ready:
-        raise HTTPException(
-            503, "Models are still loading — please retry in a few seconds"
-        )
-    qdrant = await _ensure_qdrant()
-    model = _get_embed_model()
-
-    point_id = _url_hash(body.url)
-    existing_payload = None
-    try:
-        existing = qdrant.retrieve(
-            COLLECTION_NAME,
-            ids=[point_id],
-            with_payload=True,
-            with_vectors=False,
-        )
-        if existing and existing[0].payload:
-            existing_payload = existing[0].payload
-    except Exception:
-        logger.warning(
-            "Qdrant lookup failed during index for %s — proceeding as new page",
-            body.url,
-            exc_info=True,
-        )
-
-    # Embed the content
-    loop = asyncio.get_event_loop()
-    embedding = await loop.run_in_executor(
-        None,
-        lambda: model.encode(body.content[:2000], normalize_embeddings=True).tolist(),
-    )
-
-    # Build enriched payload
-    payload = _build_index_payload(body.url, body.title, existing_payload)
-
-    # Build vector dict: at minimum the active named vector
-    active_nv = _get_active_model()
-    vectors = {active_nv: embedding}
-
-    # During dual-write phase, also embed with the target model
-    if _migration["status"] == "dual_write":
-        target_name = _migration["target_model"]
-        if target_name:
-            try:
-                target_model = SentenceTransformer(target_name)
-                target_embedding = target_model.encode(
-                    body.content[:2000], normalize_embeddings=True
-                ).tolist()
-                target_nv = _named_vector_name(target_name)
-                vectors[target_nv] = target_embedding
-
-                # Update embedding_models list
-                existing_models = json.loads(payload.get("embedding_models", "[]"))
-                for nv in [active_nv, target_nv]:
-                    if nv not in existing_models:
-                        existing_models.append(nv)
-                payload["embedding_models"] = json.dumps(existing_models)
-            except Exception as e:
-                logger.warning("Dual-write embed failed for target model: %s", e)
-
-    qdrant.upsert(
-        collection_name=COLLECTION_NAME,
-        points=[
-            models.PointStruct(
-                id=point_id,
-                vector=vectors,
-                payload=payload,
-            )
-        ],
-    )
-
-    await _evict_if_needed(qdrant)
-
-    return IndexResponse(status="indexed", url_hash=point_id)
-
-
-@app.post("/index/batch", response_model=IndexBatchResponse, status_code=201)
-async def index_batch(body: IndexBatchRequest):
-    """Embed and store multiple pages in a single batch.
-
-    Ref: ADR-0030. For large crawls, replaces N per-page POST /index
-    calls with a single batch call. Embeds all content in one
-    SentenceTransformer call and upserts via Qdrant gRPC batch.
-    Best-effort: failure is logged but never propagated.
-    """
-    if not _models_ready:
-        raise HTTPException(
-            503, "Models are still loading — please retry in a few seconds"
-        )
-    qdrant = await _ensure_qdrant()
-    model = _get_embed_model()
-
-    if not body.pages:
-        return IndexBatchResponse(status="indexed", count=0)
-
-    # Batch embed all content texts in one call
-    contents = [p.content[:2000] for p in body.pages]
-    embed_start = time.time()
-    loop = asyncio.get_event_loop()
-    embeddings = await loop.run_in_executor(
-        None,
-        lambda: model.encode(contents, normalize_embeddings=True).tolist(),
-    )
-    embed_duration = time.time() - embed_start
-    METRICS.histogram(
-        "groktocrawl_index_batch_embed_duration_seconds",
-        "Batch embedding inference latency",
-    ).observe({"batch_size": str(len(contents))}, embed_duration)
-
-    # Build points with payloads
-    active_nv = _get_active_model()
-    points = []
-    for page, embedding in zip(body.pages, embeddings):
-        point_id = _url_hash(page.url)
-        existing_payload = None
-        try:
-            existing = qdrant.retrieve(
-                COLLECTION_NAME,
-                ids=[point_id],
-                with_payload=True,
-                with_vectors=False,
-            )
-            if existing and existing[0].payload:
-                existing_payload = existing[0].payload
-        except Exception:
-            pass
-
-        payload = _build_index_payload(page.url, page.title, existing_payload)
-        vectors: dict[str, list[float]] = {active_nv: embedding}
-
-        # Dual-write support: also embed with target model
-        if _migration["status"] == "dual_write":
-            target_name = _migration["target_model"]
-            if target_name:
-                try:
-                    loop = asyncio.get_event_loop()
-                    target_embedding = await loop.run_in_executor(
-                        None,
-                        lambda: (
-                            SentenceTransformer(target_name)
-                            .encode(page.content[:2000], normalize_embeddings=True)
-                            .tolist()
-                        ),
-                    )
-                    target_nv = _named_vector_name(target_name)
-                    vectors[target_nv] = target_embedding
-                    existing_models = json.loads(payload.get("embedding_models", "[]"))
-                    for nv in [active_nv, target_nv]:
-                        if nv not in existing_models:
-                            existing_models.append(nv)
-                    payload["embedding_models"] = json.dumps(existing_models)
-                except Exception as e:
-                    logger.warning(
-                        "Batch dual-write embed failed for target model: %s", e
-                    )
-
-        points.append(
-            models.PointStruct(
-                id=point_id,
-                vector=vectors,
-                payload=payload,
-            )
-        )
-
-    # Single batch upsert via Qdrant gRPC
-    qdrant.upsert(collection_name=COLLECTION_NAME, points=points)
-
-    METRICS.counter(
-        "groktocrawl_index_batch_pages_total",
-        "Total pages indexed via batch endpoint",
-    ).inc(value=len(points))
-
-    await _evict_if_needed(qdrant)
-
-    return IndexBatchResponse(status="indexed", count=len(points))
-
-
-@app.post("/search/vector", response_model=VectorSearchResponse)
-async def search_vector(body: VectorSearchRequest):
-    """Search the vector index by semantic similarity.
-
-    Phase 4: searches the active named vector. The active model
-    is determined by _get_active_model() — defaults to env var,
-    overridable via /migrate/cutover.
-    """
-    if not _models_ready:
-        raise HTTPException(
-            503, "Models are still loading — please retry in a few seconds"
-        )
-    qdrant = await _ensure_qdrant()
-    model = _get_embed_model()
-
-    loop = asyncio.get_event_loop()
-    query_embedding = await loop.run_in_executor(
-        None,
-        lambda: model.encode(body.query, normalize_embeddings=True).tolist(),
-    )
-
-    active_nv = _get_active_model()
-
-    hits = qdrant.query_points(
-        collection_name=COLLECTION_NAME,
-        query=query_embedding,
-        using=active_nv,
-        limit=body.limit,
-    ).points
-
-    results = [
-        VectorSearchResult(
-            url=h.payload.get("url", ""),
-            title=h.payload.get("title", ""),
-            score=float(h.score),
-        )
-        for h in hits
-    ]
-
-    # Fire-and-forget access tracking (Phase 3)
-    if hits:
-        asyncio.ensure_future(_track_access(qdrant, hits))
-
-    return VectorSearchResponse(results=results)
-
-
-async def _track_access(qdrant: QdrantClient, hits: list) -> None:
-    """Increment access_count and update last_accessed_at for search results."""
-    now = _now_iso()
-    try:
-        for hit in hits:
-            point_id = hit.id
-            payload = hit.payload or {}
-            current_count = int(payload.get("access_count", 0))
-            qdrant.set_payload(
-                COLLECTION_NAME,
-                points=[point_id],
-                payload={
-                    "access_count": current_count + 1,
-                    "last_accessed_at": now,
-                },
-            )
-        logger.debug("Tracked access for %d search results", len(hits))
-    except Exception:
-        logger.debug("Failed to track access for search results", exc_info=True)
-
-
-@app.delete("/index/{url_hash}")
-async def delete_index(url_hash: int):
-    """Remove a page from the vector index by URL hash."""
-    qdrant = await _ensure_qdrant()
-    qdrant.delete(
-        COLLECTION_NAME,
-        points_selector=models.PointIdsList(points=[url_hash]),
-    )
-    return {"status": "deleted"}
-
-
-@app.get("/index/stats", response_model=IndexStatsResponse)
-async def index_stats():
-    """Return index size and configuration."""
-    qdrant = await _ensure_qdrant()
-    count = qdrant.count(COLLECTION_NAME).count
-    METRICS.gauge(
-        "groktocrawl_index_docs_total", "Current document count in the vector index"
-    ).set(value=float(count))
-    return IndexStatsResponse(total_docs=count, max_docs=MAX_DOCS)
-
-
-# ── Phase 4: Model info and migration endpoints ──────────────────
-
-
-@app.get("/index/model", response_model=ModelInfoResponse)
-async def index_model():
-    """Return current embedding model config and migration state."""
-    qdrant = await _ensure_qdrant()
-    count = qdrant.count(COLLECTION_NAME).count
-    return ModelInfoResponse(
-        current_model=EMBED_MODEL_NAME,
-        current_dim=EMBED_DIM,
-        active_named_vector=_get_active_model(),
-        collection=COLLECTION_NAME,
-        total_docs=count,
-        max_docs=MAX_DOCS,
-        migration={
-            "status": _migration["status"],
-            "source_model": _migration["source_model"],
-            "target_model": _migration["target_model"],
-            "docs_processed": _migration["docs_processed"],
-            "docs_total": _migration["docs_total"],
-        },
-    )
-
-
-@app.post("/index/migrate/start", status_code=202)
-async def migrate_start(body: MigrationStartRequest):
-    """Start an embedding model migration (backfill phase).
-
-    Background task scrolls all points, re-embeds with the target
-    model, and adds a named vector. Queries continue using the
-    active model until cutover.
-    """
-    global _migration_task, _migration
-
-    if _migration["status"] != "idle":
-        raise HTTPException(
-            409, f"Migration already in progress: {_migration['status']}"
-        )
-
-    qdrant = await _ensure_qdrant()
-
-    target_model_id = body.target_model
-    target_dim = body.target_dim
-
-    _migration = {
-        "status": "backfilling",
-        "source_model": EMBED_MODEL_NAME,
-        "source_dim": EMBED_DIM,
-        "target_model": target_model_id,
-        "target_dim": target_dim,
-        "docs_processed": 0,
-        "docs_total": 0,
-        "started_at": _now_iso(),
-        "completed_at": "",
-    }
-
-    _migration_task = app.state.task_tracker.create_background_task(
-        _run_backfill(qdrant, target_model_id, target_dim)
-    )
-
-    return {
-        "status": "accepted",
-        "message": f"Backfill started: {EMBED_MODEL_NAME} -> {target_model_id}",
-    }
-
-
-@app.get("/index/migrate/status", response_model=MigrationStatusResponse)
-async def migrate_status():
-    """Return migration progress."""
-    return MigrationStatusResponse(
-        status=_migration["status"],
-        source_model=_migration["source_model"],
-        source_dim=_migration["source_dim"],
-        target_model=_migration["target_model"],
-        target_dim=_migration["target_dim"],
-        docs_processed=_migration["docs_processed"],
-        docs_total=_migration["docs_total"],
-        started_at=_migration["started_at"],
-        completed_at=_migration["completed_at"],
-    )
-
-
-@app.post("/index/migrate/cutover")
-async def migrate_cutover():
-    """Switch queries to the new model.
-
-    The in-memory active model override flips to the target model's
-    named vector. On container restart, reverts to ACTIVE_EMBED_MODEL
-    env var. For permanent cutover, update the env var and restart.
-    """
-    global _active_override
-
-    if _migration["status"] not in ("dual_write", "backfilling"):
-        raise HTTPException(
-            409, f"Cannot cutover: migration status is '{_migration['status']}'"
-        )
-
-    if not _migration["target_model"]:
-        raise HTTPException(400, "No target model configured")
-
-    target_nv = _named_vector_name(_migration["target_model"])
-    _active_override = target_nv
-    _migration["status"] = "cutover"
-    _migration["completed_at"] = _now_iso()
-
-    logger.info(
-        "Migration cutover: active model switched to '%s' (named vector: '%s')",
-        _migration["target_model"],
-        target_nv,
-    )
-
-    return {
-        "status": "cutover",
-        "active_named_vector": target_nv,
-        "message": f"Queries now using '{target_nv}'. "
-        f"Update ACTIVE_EMBED_MODEL env var and restart to persist.",
-    }
-
-
 # ── Metrics endpoint ──────────────────────────────────────────────
 
 
@@ -1250,3 +395,16 @@ async def metrics():
         content=text,
         media_type="application/openmetrics-text; version=1.0.0",
     )
+
+
+# ── Router includes ─────────────────────────────────────────────────
+# Import routers at module bottom to avoid circular imports.
+# All module-level symbols needed by routers are defined above.
+
+from router_index import router_index  # noqa: E402
+from router_migration import router_migration  # noqa: E402
+from router_search import router_search  # noqa: E402
+
+app.include_router(router_index, prefix="/index")
+app.include_router(router_search, prefix="/search")
+app.include_router(router_migration, prefix="/index/migrate")
